+++ haarpy/tests/test_unitary.py
@@ -19,45 +19,25 @@ import pytest
 import haarpy as ap
 from fractions import Fraction
 from sympy.combinatorics import Permutation
-from numpy import allclose
-from sympy import Symbol, simplify
+from sympy import Symbol, simplify, fraction, factor
 from sympy.combinatorics.named_groups import SymmetricGroup
 
 
 @pytest.mark.parametrize(
-    "degree,cycle,conjugacy",
+    "degree, cycle, conjugacy",
     [
         (3, Permutation(0, 1, 2), (3,)),
-        (
-            3,
-            Permutation(
-                2,
-            )(0, 1),
-            (2, 1),
-        ),
+        (3, Permutation(2)(0, 1), (2, 1)),
         (4, Permutation(0, 1, 2, 3), (4,)),
         (4, Permutation(0, 1, 2), (3, 1)),
         (4, Permutation(0, 1)(2, 3), (2, 2)),
         (5, Permutation(0, 1, 2), (3, 1, 1)),
         (5, Permutation(1, 2, 4), (3, 1, 1)),
-        (
-            5,
-            Permutation(
-                3,
-            )(0, 1, 2),
-            (3, 1, 1),
-        ),
-        (
-            6,
-            Permutation(
-                2,
-            )(
-                3, 4, 5
-            )(0, 1),
-            (3, 2, 1),
-        ),
+        (5, Permutation(3)(0, 1, 2), (3, 1, 1)),
+        (6, Permutation(2)(3, 4, 5)(0, 1), (3, 2, 1)),
         (6, Permutation(2, 3, 4, 0, 1, 5), (6,)),
         (7, Permutation(2, 3, 4, 0, 1, 6), (6, 1)),
+        (1, Permutation(0), (1,)),
     ],
 )
 def test_get_class(degree, cycle, conjugacy):
@@ -66,7 +46,7 @@ def test_get_class(degree, cycle, conjugacy):
 
 
 @pytest.mark.parametrize(
-    "degree,cycle",
+    "degree, cycle",
     [
         ("a", Permutation(0, 1, 2)),
         (0.1, Permutation(2)(0, 1)),
@@ -80,18 +60,18 @@ def test_get_class_degree_type_error(degree, cycle):
         ap.get_class(cycle, degree)
 
 
-@pytest.mark.parametrize("degree", range(-2, 2))
+@pytest.mark.parametrize("degree", range(-3, 1))
 def test_get_class_degree_value_error(degree):
     """Test the degree parameter ValueError"""
     with pytest.raises(
         ValueError,
-        match=".*The degree you have provided is too low. It must be an integer greater than 1.*",
+        match=".*The degree you have provided is too low. It must be an integer greater than 0.*",
     ):
         ap.get_class(Permutation(0, 1, 2), degree)
 
 
 @pytest.mark.parametrize(
-    "degree,cycle",
+    "degree, cycle",
     [
         (3, [1, 2, 3]),
         (4, "a"),
@@ -108,7 +88,7 @@ def test_get_class_cycle_type_error(degree, cycle):
 
 
 @pytest.mark.parametrize(
-    "degree,cycle",
+    "degree, cycle",
     [
         (3, Permutation(0, 1, 2, 3)),
         (3, Permutation(2, 3)(0, 1)),
@@ -118,16 +98,16 @@ def test_get_class_cycle_type_error(degree, cycle):
 )
 def test_get_class_cycle_value_error(degree, cycle):
     """Test the cycle parameter ValueError if permutation maximum value is greater than the degree"""
-    with pytest.raises(
-        ValueError, match=".*Incompatible degree and permutation cycle.*"
-    ):
+    with pytest.raises(ValueError, match=".*Incompatible degree and permutation cycle.*"):
         ap.get_class(cycle, degree)
 
 
 @pytest.mark.parametrize(
-    "tableau,add_unit,partition,result",
+    "tableau, add_unit,partition, result",
     [
         ([[0, 0, 0], [0, 0, 0]], 1, (3, 1), [[[1, 0, 0], [0, 0, 0]]]),
+        ([[1, 2, 0], [2, 3, 0]], 4, [3, 2], [[[1, 2, 4], [2, 3, 0]]]),
+        ([[1, 1, 1, 3], [2, 0, 0, 0]], 4, (4, 2), [[[1, 1, 1, 3], [2, 4, 0, 0]]]),
         (
             [[1, 0, 0], [0, 0, 0]],
             1,
@@ -140,14 +120,12 @@ def test_get_class_cycle_value_error(degree, cycle):
             (3, 2),
             [[[1, 2, 2], [0, 0, 0]], [[1, 2, 0], [2, 0, 0]]],
         ),
-        ([[1, 2, 0], [2, 3, 0]], 4, [3, 2], [[[1, 2, 4], [2, 3, 0]]]),
         (
             [[0, 0, 0], [0, 0, 0], [0, 0, 0]],
             1,
             (3, 1, 1),
             [[[1, 0, 0], [0, 0, 0], [0, 0, 0]]],
         ),
-        ([[1, 1, 1, 3], [2, 0, 0, 0]], 4, (4, 2), [[[1, 1, 1, 3], [2, 4, 0, 0]]]),
         (
             [[1, 2, 3, 0], [1, 2, 0, 0], [1, 0, 0, 0]],
             3,
@@ -225,11 +203,12 @@ def test_ssyt(partition, conjugacy_class, result):
         ([[1, 1, 2, 2], [3, 3, 0, 0]], (2, 2, 2)),
         ([[1, 1, 3, 3], [2, 2, 0, 0]], (2, 2, 2)),
         ([[1, 2, 3, 3], [1, 2, 0, 0]], (2, 2, 2)),
+        ([[1, 1, 1, 1, 1], [1, 2, 2, 0, 0]], (6, 2)),
     ],
 )
-def test_bad_mapping_false(tableau, conjugacy_class):
+def test_border_strip_tableau_true(tableau, conjugacy_class):
     "Test bad_mapping for well mapped tableaux, returns False"
-    assert not ap.bad_mapping(tableau, conjugacy_class)
+    assert ap.border_strip_tableau(tableau, conjugacy_class)
 
 
 @pytest.mark.parametrize(
@@ -241,11 +220,13 @@ def test_bad_mapping_false(tableau, conjugacy_class):
         ([[1, 2, 4], [2, 3, 0]], (1, 2, 1, 1)),
         ([[1, 1, 2, 3], [2, 3, 0, 0]], (2, 2, 2)),
         ([[1, 2, 2, 3], [1, 3, 0, 0]], (2, 2, 2)),
+        ([[1, 1, 1, 1, 1], [1, 1, 2, 0, 0]], (7, 1)),
+        ([[1, 1, 1, 1, 2], [1, 1, 1, 0, 0]], (7, 1)),
     ],
 )
-def test_bad_mapping_true(tableau, conjugacy_class):
+def test_border_strip_tableau_false(tableau, conjugacy_class):
     "Test bad_mapping for wrong mapped tableaux, returns True"
-    assert ap.bad_mapping(tableau, conjugacy_class)
+    assert not ap.border_strip_tableau(tableau, conjugacy_class)
 
 
 @pytest.mark.parametrize(
@@ -272,11 +253,15 @@ def test_bad_mapping_true(tableau, conjugacy_class):
         ((4, 3), (1, 1, 1, 1, 1, 1, 1), 14),
         ((2, 2, 1), (2, 3), -1),
         ((2, 2, 2), (4, 2), -1),
+        ((5, 3), (5, 1, 1, 1), -2),
+        ((5, 3), (6, 1, 1), -1),
+        ((5, 3), (7, 1), 0),
+        ((4, 2, 1, 1), (7, 1), -1),
     ],
 )
-def test_murn_naka(partition, conjugacy_class, caracter):
-    "Test murn_naka based on the outputs form weingarten mathematica package"
-    assert ap.murn_naka(partition, conjugacy_class) == caracter
+def test_murn_naka_rule(partition, conjugacy_class, caracter):
+    "Test murn_naka_rule based on the outputs form weingarten mathematica package"
+    assert ap.murn_naka_rule(partition, conjugacy_class) == caracter
 
 
 @pytest.mark.parametrize(
@@ -311,10 +296,10 @@ def test_sn_dimension(partition, dimension):
         ((4, 3, 2, 1, 1)),
     ],
 )
-def test_sn_dimension_murn_naka(partition):
-    "Reconcil sn_dimension and murn_naka for a class mu of ones"
+def test_sn_dimension_murn_naka_rule(partition):
+    "Reconcil sn_dimension and murn_naka_rule for a class mu of ones"
     mu = sum(partition) * [1]
-    assert ap.sn_dimension(partition) == ap.murn_naka(partition, mu)
+    assert ap.sn_dimension(partition) == ap.murn_naka_rule(partition, mu)
 
 
 @pytest.mark.parametrize(
@@ -353,48 +338,49 @@ def test_ud_dimension_wrong_dimension(partition):
 
 
 @pytest.mark.parametrize(
-    "conjugacy, num, denum",
+    "conjugacy, dimension, num, denum",
     [
-        ((1,), 1, 7),
-        ((2,), -1, 336),
-        ((1, 1), 1, 48),
-        ((2, 1), -1, 2160),
-        ((1, 1, 1), 47, 15120),
-        ((3, 1), 19, 846720),
-        ((2, 2), 11, 846720),
-        ((3, 2), -61, 69854400),
-        ((3, 1, 1), 1, 249480),
-        ((2, 2, 1), 47, 19958400),
-        ((3, 3), 311, 3353011200),
-        ((2, 1, 1, 1, 1), -421, 191600640),
-        ((1, 1, 1, 1, 1, 1), 82477, 6706022400),
-        ((4, 3), -17, 792529920),
+        ((1,), 7, 1, 7),
+        ((2,), 7, -1, 336),
+        ((1, 1), 7, 1, 48),
+        ((2, 1), 7, -1, 2160),
+        ((1, 1, 1), 7, 47, 15120),
+        ((3, 1), 7, 19, 846720),
+        ((2, 2), 7, 11, 846720),
+        ((3, 2), 7, -61, 69854400),
+        ((3, 1, 1), 7, 1, 249480),
+        ((2, 2, 1), 7, 47, 19958400),
+        ((3, 3), 7, 311, 3353011200),
+        ((2, 1, 1, 1, 1), 7, -421, 191600640),
+        ((1, 1, 1, 1, 1, 1), 7, 82477, 6706022400),
+        ((4, 3), 7, -17, 792529920),
+        ((7, 1), 8, 151, 317011968000),
     ],
 )
-def test_weingarten_class(conjugacy, num, denum):
+def test_weingarten_class(conjugacy, dimension, num, denum):
     "Test weingarten_class based on the outputs form weingarten mathematica package"
-    assert ap.weingarten_class(conjugacy, 7) == Fraction(num, denum)
+    assert ap.weingarten_class(conjugacy, dimension) == Fraction(num, denum)
 
 
 @pytest.mark.parametrize(
-    "cycle, degree, num, denum",
+    "cycle, degree, dimension, num, denum",
     [
-        (Permutation(0, 1), 2, -1, 336),
-        (Permutation(0), 2, 1, 48),
-        (Permutation(0, 1, 2), 3, 1, 7560),
-        (Permutation(1, 2), 3, -1, 2160),
-        (Permutation(2), 3, 47, 15120),
-        (Permutation(0, 1, 2), 4, 19, 846720),
-        (Permutation(0, 2)(1, 3), 4, 11, 846720),
-        (Permutation(2), 4, 403, 846720),
-        (Permutation(4, 1), 5, -1739, 139708800),
-        (Permutation(2), 5, 1499, 19958400),
-        (Permutation(3, 4, 5), 6, 5167, 6706022400),
+        (Permutation(0, 1), 2, 7, -1, 336),
+        (Permutation(0), 2, 7, 1, 48),
+        (Permutation(0, 1, 2), 3, 7, 1, 7560),
+        (Permutation(1, 2), 3, 7, -1, 2160),
+        (Permutation(2), 3, 7, 47, 15120),
+        (Permutation(0, 1, 2), 4, 7, 19, 846720),
+        (Permutation(0, 2)(1, 3), 4, 7, 11, 846720),
+        (Permutation(2), 4, 7, 403, 846720),
+        (Permutation(4, 1), 5, 7, -1739, 139708800),
+        (Permutation(2), 5, 7, 1499, 19958400),
+        (Permutation(3, 4, 5), 6, 7, 5167, 6706022400),
     ],
 )
-def test_weingarten_element(cycle, degree, num, denum):
+def test_weingarten_element(cycle, degree, dimension, num, denum):
     "Test weingarten_element based on the outputs form weingarten mathematica package"
-    assert ap.weingarten_element(cycle, degree, 7) == Fraction(num, denum)
+    assert ap.weingarten_element(cycle, degree, dimension) == Fraction(num, denum)
 
 
 @pytest.mark.parametrize(
@@ -453,3 +439,47 @@ def test_gram_orthogonality(n):
     weingarten = lambda g, d, n: ap.weingarten_element(g, n, d)
     orthogonality = sum(character(g, d, n) * weingarten(g, d, n) for g in group(n))
     assert simplify(orthogonality) == 1
+
+
+@pytest.mark.parametrize(
+    "sequences, weingarten_map",
+    [
+        (((1,), (2,), (1,), (2,)), {(1,): 1}),
+        (("ik", "jj", "ik", "jj"), {(1, 1): 1, (2,): 1}),
+        (((1, 3), (2, 2), (1, 3), (2, 2)), {(1, 1): 1, (2,): 1}),
+        (("ik", "jl", "ik", "jl"), {(1, 1): 1}),
+        (("ik", "jl", "ik", "lj"), {(2,): 1}),
+        (("ikm", "jln", "ikm", "jln"), {(1, 1, 1): 1}),
+        (("ijk", "lmn", "ijk", "mnl"), {(3,): 1}),
+        (("ijk", "llm", "ijk", "lml"), {(3,): 1, (2, 1): 1}),
+        (((1, 2, 3), (4, 4, 4), (1, 2, 3), (4, 4, 4)), {(1, 1, 1): 1, (2, 1): 3, (3,): 2}),
+        (((1, 1, 3), (1, 4, 4), (1, 2, 3), (1, 4, 4)), {(3,): 0}),
+        (((1, 2, 3), (4, 4, 4), (1, 2, 4), (4, 4, 4)), {(3,): 0}),
+        (((1, 2, 3), (4, 4, 4), (1, 2), (4, 4)), {(3,): 0}),
+    ],
+)
+def test_haar_integral_hand(sequences, weingarten_map):
+    "Test integral of Haar distribution unitaries against hand-calculated integrals"
+    dimension = Symbol("d")
+    integral = sum(
+        frequency * ap.weingarten_class(conjugacy, dimension)
+        for conjugacy, frequency in weingarten_map.items()
+    )
+    numerator, denominator = fraction(simplify(integral))
+    integral = factor(numerator) / factor(denominator)
+    assert ap.haar_integral(sequences, dimension) == integral
+
+
+@pytest.mark.parametrize(
+    "sequence",
+    [
+        ((1,), (1,), (1,)),
+        ((1, 1, 1), (1, 1), (1, 1, 1), (1, 1, 1)),
+        ((1, 1, 1), (1, 1, 1), (1, 1), (1, 1, 1)),
+    ],
+)
+def test_haar_integral_wrong_format(sequence):
+    """Test wrong tuple format ValueError"""
+    dimension = Symbol("d")
+    with pytest.raises(ValueError, match="Wrong tuple format"):
+        ap.haar_integral(sequence, dimension)